[REF] Explode 'misc' to several 'utils.*' modules
